+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - All clients now support a user-provided [httpx.Client](https://www.python-httpx.org/api/#client)
   objects.
+- Add support for the VBML `absolutePosition` style.
 
 ### Changed
 - The `encode_*()` functions now consistently specify keyword-only arguments.